requirements_test.txt
requests
replicate
pytest
Pillow
numpy